Update tut-making-platformer.md
You named the crystal widget "CrystalCounter" in the tutorial, not "CrystalsWidget". 
(also, the line "ct.types.copy('CrystalCounter', 0, 0);" does not appear on the screenshot below, and is not explained)

In the last code snippet (better code for moving platforms), you look for the robot two times (one at this.y and another one at this.y -1). Is there a reason for that ?

Thanks for the tutorial ;)